api/requirements/base.txt
# Tools
pre-commit

# Django
Django==2.2.7
django-environ==0.4.5
django-redis==4.12.1
requests==2.23.0

# Python-PostgreSQL Database Adapter
psycopg2-binary==2.8.4

# Models
django-model-utils==3.2.0
python_dateutil==2.8.0

# Utils
django-ipware==2.1.0

# DRF
djangorestframework==3.10.3
django-oauth-toolkit==1.2.0
django-cors-middleware==1.4.0
django-rest-auth==0.9.5
django-allauth==0.40.0
django-filter==2.4.0


# Document
drf-yasg==1.17.0

# Celery
celery==4.3.0
#eventlet==0.25.1
gevent==21.12.0
redis==3.3.11
django-celery-beat==1.5.0
django-celery-results==1.1.2
tzdata==2022.1


# debug
django_extensions==2.2.5
django-debug-toolbar==2.1

# WSGI Handler
gunicorn==20.0.2

# code coverage and testing
django-webtest==1.9.7
coverage==4.5.4

# storage
boto3==1.10.26
django-storages==1.8


# JWT
PyJWT==1.7.1
djangorestframework_simplejwt==4.4.0
cryptography==3.1

# Bug Service
sentry-sdk==1.5.2
bugsnag==3.6.0


# (Mime type) Upload app
python_magic==0.4.15

# Social core
social-auth-core==3.3.3


# Stripe
stripe==2.48.0

# Fix celery
vine==1.3.0

# Caching
django-cache-memoize==0.1.8

# zipcode
zipcodes==1.1.3

# authentication using aws cognito service 
django-cognito-jwt

# filter list
django-filter

